requirements.txt
# YouTube Shorts Automation Pipeline Dependencies

# AI/LLM
openai>=1.0.0

# Image Processing
Pillow>=10.0.0

# Video Processing
moviepy>=1.0.3

# Audio Processing
pydub>=0.25.1

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
tqdm>=4.66.0

